Move KuudraTier to KuudraUtils.kt
